:app — split Data sources into Location + Calendar; cache last device fix
Move location into its own settings sub-page and calendar into a separate
one. Replaces the bundled "Data sources" route with two adjacent routes
(Location, Calendar); the existing Today "set up location" deep-link now
targets Location. The remembered route survives the rename via a custom
Saver that falls back to Root if an unknown enum name comes back from
process-death restore.

Redesign the location permission UX so background-grant is the visible
primary action instead of being buried below the toggle: a warning
banner ("Always-on location needed — Grant") sits at the top of the
Location page whenever device-location is on but ACCESS_BACKGROUND_LOCATION
isn't granted, and a smaller deep-link card surfaces the same state from
the Settings root. Trim the wordy strings (settings_location_description*,
_background_*body, _calendar_description_*) to a single short sentence
each so the page scans cleanly. Drop the in-toggle filled "Grant
always-on" button — the banner now carries that CTA.

Cache the most recent device-resolved location so users don't need to
manually pick a city as a safety net. The worker now writes through to
SettingsRepository.setLocation() after every successful device-resolve;
when the next worker run can't read the device fix (provider blip,
services briefly off) it falls through to the cached value transparently.
To populate the cache without waiting for tomorrow's run, toggling
device-location ON enqueues a one-shot FetchAndNotifyWorker so the user
sees their city in Settings within seconds. Manual entry is now the
explicit "the system returned the wrong location" override — picking a
city via the search dialog auto-flips useDeviceLocation off so the next
worker run doesn't immediately overwrite the manual pick.

Privacy: no change to what leaves the device. Forecast/geocoding
requests, the rendered insight, and TTS payloads are unchanged.
