+++ app/src/main/res/values-pt-rBR/strings.xml
@@ -55,7 +55,6 @@ zh-CN) is unchanged; only the displayed label localizes.
     <string name="settings_root_schedule">Agendamento</string>
     <string name="settings_root_region">Região</string>
     <string name="settings_root_clothes">Roupas</string>
-    <string name="settings_root_data_sources">Fontes de dados</string>
     <string name="settings_root_about">Sobre</string>
 
     <string name="garment_sweater">Suéter</string>
@@ -295,10 +294,6 @@ zh-CN) is unchanged; only the displayed label localizes.
     <string name="settings_location_use_device">Usar localização do aparelho</string>
     <string name="settings_location_grant_background">Conceder localização em segundo plano (configurações do sistema)</string>
     <string name="settings_location_unset">Localização não definida</string>
-    <string name="settings_location_description">A previsão é buscada para este local. Pesquise pelo nome da cidade; os resultados vêm do serviço de geocodificação gratuito da Open-Meteo.</string>
-    <string name="settings_location_description_device_on">Quando ativada, ClothesCast lê a localização aproximada do seu aparelho no momento da notificação (apenas torre celular / WiFi — sem GPS por hardware). A localização abaixo é usada como reserva se a leitura do aparelho falhar. A localização em segundo plano precisa estar concedida nas configurações do sistema, senão o Worker matinal não consegue lê-la.</string>
-    <string name="settings_location_set">Definir localização</string>
-    <string name="settings_location_change">Alterar localização</string>
     <string name="settings_location_clear">Limpar localização</string>
     <string name="settings_location_search_title">Pesquisar uma localização</string>
     <string name="settings_location_query_label">Cidade ou lugar</string>
@@ -531,7 +526,6 @@ zh-CN) is unchanged; only the displayed label localizes.
     <string name="settings_root_clothes_subtitle">Temperaturas e roupas</string>
     <string name="settings_root_region_subtitle">Idiomas e unidades</string>
     <string name="settings_root_voice_subtitle">Provedores e sotaques</string>
-    <string name="settings_root_data_sources_subtitle">Localização e calendário</string>
 
     <string name="settings_tts_device_voice_label">Voz do dispositivo</string>
     <string name="settings_tts_device_voice_auto">Automático (melhor para sotaque)</string>
