+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -46,7 +46,6 @@ label localizes.
     <string name="settings_root_schedule">计划</string>
     <string name="settings_root_region">地区</string>
     <string name="settings_root_clothes">服装</string>
-    <string name="settings_root_data_sources">数据源</string>
     <string name="settings_root_about">关于</string>
 
     <string name="garment_sweater">毛衣</string>
@@ -286,10 +285,6 @@ label localizes.
     <string name="settings_location_use_device">使用设备位置</string>
     <string name="settings_location_grant_background">授予后台位置权限（系统设置）</string>
     <string name="settings_location_unset">未设置位置</string>
-    <string name="settings_location_description">天气预报是针对这个地点获取的。按城市名称搜索；结果来自 Open-Meteo 的免费地理编码服务。</string>
-    <string name="settings_location_description_device_on">启用后，ClothesCast 在通知时会读取设备的大致位置（仅基站 / WiFi — 不使用 GPS 硬件定位）。如果设备读取失败，会使用下方的位置作为备用。后台位置权限必须在系统设置里授予，否则早晨的 Worker 无法读取。</string>
-    <string name="settings_location_set">设置位置</string>
-    <string name="settings_location_change">更改位置</string>
     <string name="settings_location_clear">清除位置</string>
     <string name="settings_location_search_title">搜索位置</string>
     <string name="settings_location_query_label">城市或地点</string>
@@ -521,7 +516,6 @@ label localizes.
     <string name="settings_root_clothes_subtitle">温度与穿搭</string>
     <string name="settings_root_region_subtitle">语言与单位</string>
     <string name="settings_root_voice_subtitle">提供商与口音</string>
-    <string name="settings_root_data_sources_subtitle">位置与日历</string>
 
     <string name="settings_tts_device_voice_label">设备语音</string>
     <string name="settings_tts_device_voice_auto">自动（最适合口音）</string>
